IMplementing bigdecimal

+++ java/src/main/java/com/techelevator/models/Candy.java
@@ -1,8 +1,10 @@
 package com.techelevator.models;
 
+import java.math.BigDecimal;
+
 public class Candy extends Items {
 
-    public Candy(String name, double price, String slotNumber, String dispenseMessage) {
+    public Candy(String name, BigDecimal price, String slotNumber, String dispenseMessage) {
         super(name, price, slotNumber, dispenseMessage);
     }
 }
