+++ java/src/main/java/com/techelevator/models/Drink.java
@@ -1,8 +1,10 @@
 package com.techelevator.models;
 
+import java.math.BigDecimal;
+
 public class Drink extends Items {
 
-    public Drink(String name, double price, String slotNumber, String dispenseMessage) {
+    public Drink(String name, BigDecimal price, String slotNumber, String dispenseMessage) {
         super(name, price, slotNumber, dispenseMessage);
     }
