+++ java/src/main/java/com/techelevator/models/Gum.java
@@ -1,8 +1,10 @@
 package com.techelevator.models;
 
+import java.math.BigDecimal;
+
 public class Gum extends Items {
 
-    public Gum(String name, double price, String slotNumber, String dispenseMessage) {
+    public Gum(String name, BigDecimal price, String slotNumber, String dispenseMessage) {
         super(name, price, slotNumber, dispenseMessage);
     }
