+++ java/src/main/java/com/techelevator/models/Munchy.java
@@ -1,8 +1,10 @@
 package com.techelevator.models;
 
+import java.math.BigDecimal;
+
 public class Munchy extends Items{
 
-    public Munchy(String name, double price, String slotNumber, String dispenseMessage) {
+    public Munchy(String name, BigDecimal price, String slotNumber, String dispenseMessage) {
         super(name, price, slotNumber, dispenseMessage);
     }
